+++ packages/localize-tools/testdata/build-runtime-xliff/goldens/foo.ts
@@ -45,5 +45,4 @@ msg(html`a:${'A'} b:${'B'} c:${'C'}`);
 // Custom ID
 msg('Hello World', {id: 'myId'});
 
-/** @desc Description of 0 */
-msg('described 0');
+msg('described 0', {desc: 'Description of 0'});